refactor: move telemetry functions outside codewhispererServer.ts (#1983)
* refactor: move telemetry functions outside codewhispererServer.ts

* refactor: patch
